Fixing InterpreterSpec
